+++ constrain/Classes/Constraints.swift
@@ -4,27 +4,7 @@
 
 import UIKit
 
-public enum ConstraintIdentifier: String {
-    case top = ".top"
-    case leading = ".leading"
-    case trailing = ".trailing"
-    case bottom = ".bottom"
-    case topSafe = ".topSafe"
-    case bottomSafe = ".bottomSafe"
-    case height = ".height"
-    case width = ".width"
-    case centerX = ".centerX"
-    case centerY = ".centerY"
-    case aspectRatio = ".aspectRatio"
-
-    // Not actually constraints:
-    case cornerRadius = ".cornerRadius"
-    case size = ".size" // size combines .width and .height
-}
-
 public typealias Relationship = NSLayoutConstraint.Relation
-
-public typealias ConstraintSet = [NSLayoutConstraint]
 
 /// Constraints is a container of layout constraints for a UIView. It has convenience methods for creating and retrieving constraints.
 /// For improved readability, all methods can be chained.
@@ -134,75 +114,4 @@ extension Constraints {
         allConstraints.append(constraint)
         latestConstraint = constraint
     }
-}
-
-extension Constraints {
-    
-    /// When storing a reference to a Constraints instance this method allows to retrieve a respective constraint.
-    public func layoutConstraintWithIdentifier(_ identifier: ConstraintIdentifier) -> NSLayoutConstraint? {
-        return constraints[identifier]
-    }
-    
-    /// When storing a reference to a Constraints instance this method allows to set the constant of a respective constraint.
-    public func setConstant(_ constant: CGFloat, forIdentifier identifier: ConstraintIdentifier) {
-        layoutConstraintWithIdentifier(identifier)?.constant = constant
-    }
-
-    @discardableResult
-    public func update(_ identifier: ConstraintIdentifier, to constant: CGFloat) -> Self {
-        // constrain will start to support setting and updating of view propperties that are not actually NSLayoutConstraints. We have to treat them slightly differently.
-        switch identifier {
-        case .cornerRadius:
-            cornerRadius(constant)
-        case .size:
-            setConstant(constant, forIdentifier: .width)
-            setConstant(constant, forIdentifier: .height)
-        default:
-            setConstant(constant, forIdentifier: identifier)
-        }
-        return self
-    }
-
-    @discardableResult
-    public func cornerRadius(_ value: CGFloat) -> Self {
-        view?.layer.cornerRadius = value
-        return self
-    }
-    
-}
-    
-extension Constraints {
-    
-    @discardableResult
-    public func refresh() -> Self {
-        view?.setNeedsLayout()
-        view?.layoutIfNeeded()
-        return self
-    }
-    
-}
-
-
-// bulk activation/deactivation
-extension Constraints {
-    
-    @discardableResult
-    public func activate() -> Self {
-        NSLayoutConstraint.activate(allConstraints)
-        isActive = true
-        return self
-    }
-    
-    @discardableResult
-    public func deactivate() -> Self {
-        NSLayoutConstraint.deactivate(allConstraints)
-        isActive = false
-        return self
-    }
-    
-    @discardableResult
-    public func toggle() -> Self {
-        return isActive ? deactivate() : activate()
-    }
-    
 }
